Feature 4/requirements.txt
streamlit==1.33.0
google-generativeai==0.3.2
python-dotenv==1.0.1
pandas==2.2.1
plotly==5.18.0